+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [0.3.1(-bc)] - 2025-04-29
+
+### Added
+
+- `generic-array-struct-builder` crate
+
+
+## [0.3.0-bc] - 2025-03-24
 
 ### Breaking
 
-- create older rustc version compatible versions of the crate behind a `-bc` pre-release version
+- create older-rustc-version-compatible versions of the crate behind a `-bc` pre-release version
+
 
 ## [0.2.0] - 2025-03-23
 
@@ -20,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `*_mut(&mut self) -> &mut T` methods
+
 
 ## [0.1.0] - 2025-03-22